+++ src/components/buytokens/TokensReceipt.vue
