+++ src/pages/Discover.vue
